+++ src/knapsack.rs
@@ -1,15 +1,15 @@
 #[derive(Debug)]
 pub struct KnapsackProblem {
-    pub n_items: u16,
-    pub capacity: u16,
+    pub n_items: usize,
+    pub capacity: usize,
     pub treasure_items: Vec<KnapsackItem>,
 }
 
 impl KnapsackProblem {
-    pub fn new(n_items: u16, capacity: u16, treasure_items: Vec<KnapsackItem>) -> KnapsackProblem {
+    pub fn new(n_items: usize, capacity: usize, treasure_items: Vec<KnapsackItem>) -> KnapsackProblem {
         assert_eq!(
             treasure_items.len(),
-            n_items as usize,
+            n_items,
             "Size of treasure items does not match n_items"
         );
         KnapsackProblem {
@@ -21,6 +21,6 @@ impl KnapsackProblem {
 }
 #[derive(Debug)]
 pub struct KnapsackItem {
-    pub value: u16,
-    pub weight: u16,
+    pub value: usize,
+    pub weight: usize,
 }
